+++ tests/src/Functional/SmsFrameworkBrowserTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\sms\Tests;
+namespace Drupal\Tests\sms\Functional;
 
 use Drupal\Core\Url;
 use Drupal\sms\Entity\SmsMessage;
@@ -12,7 +12,7 @@ use Drupal\sms\Direction;
  * @group SMS Framework
  * @group legacy
  */
-class SmsFrameworkWebTest extends SmsFrameworkWebTestBase {
+class SmsFrameworkBrowserTest extends SmsFrameworkBrowserTestBase {
 
   /**
    * Tests queue statistics located on Drupal report page.
@@ -45,8 +45,8 @@ class SmsFrameworkWebTest extends SmsFrameworkWebTestBase {
     $this->drupalLogin($this->rootUser);
     $this->drupalGet(Url::fromRoute('system.status'));
 
-    $this->assertRaw('There are 2 messages in the incoming queue.');
-    $this->assertRaw('There are 4 messages in the outgoing queue.');
+    $this->assertSession()->responseContains('There are 2 messages in the incoming queue.');
+    $this->assertSession()->responseContains('There are 4 messages in the outgoing queue.');
   }
 
 }
