+++ tests/src/Functional/SmsFrameworkDeliveryReportTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\sms\Tests;
+namespace Drupal\Tests\sms\Functional;
 
 use Drupal\Core\Url;
 use Drupal\sms\Message\SmsDeliveryReportInterface;
@@ -13,7 +13,7 @@ use Drupal\sms\Message\SmsMessageResultInterface;
  * @group SMS Framework
  * @group legacy
  */
-class SmsFrameworkDeliveryReportTest extends SmsFrameworkWebTestBase {
+class SmsFrameworkDeliveryReportTest extends SmsFrameworkBrowserTestBase {
 
   /**
    * Tests delivery reports integration.
@@ -31,14 +31,14 @@ class SmsFrameworkDeliveryReportTest extends SmsFrameworkWebTestBase {
 
     $result = $sms_messages[0]->getResult();
     $this->assertTrue($result instanceof SmsMessageResultInterface);
-    $this->assertEqual(count($sms_message->getRecipients()), count($result->getReports()));
+    $this->assertEquals(count($sms_message->getRecipients()), count($result->getReports()));
     $reports = $result->getReports();
 
     /** @var \Drupal\sms\Message\SmsDeliveryReportInterface $first_report */
     $first_report = reset($reports);
     $message_id = $first_report->getMessageId();
     $this->assertTrue($first_report instanceof SmsDeliveryReportInterface);
-    $this->assertEqual($first_report->getStatus(), SmsMessageReportStatus::QUEUED);
+    $this->assertEquals($first_report->getStatus(), SmsMessageReportStatus::QUEUED);
 
     // Get the delivery reports url and simulate push delivery report.
     $url = $test_gateway->getPushReportUrl()->setAbsolute()->toString();
@@ -56,16 +56,19 @@ class SmsFrameworkDeliveryReportTest extends SmsFrameworkWebTestBase {
    ]
 }
 EOF;
-    $this->drupalPost($url, 'application/json', ['delivery_report' => $delivery_report]);
-    $this->assertText('custom response content');
+    $client = $this->getSession()->getDriver();
+//    $client = $this->getHttpClient();
+//    return $client->request($method, $url->setAbsolute(TRUE)->toString(), $request_options);
+//    $this->drupalPost($url, 'application/json', ['delivery_report' => $delivery_report]);
+    $this->assertSession()->pageTextContains('custom response content');
     \Drupal::state()->resetCache();
 
     // Get the stored report and verify that it was properly parsed.
     $second_report = $this->getTestMessageReport($message_id, $test_gateway);
     $this->assertTrue($second_report instanceof SmsDeliveryReportInterface);
-    $this->assertEqual("status message", $second_report->getStatusMessage());
-    $this->assertEqual($delivered_time, $second_report->getTimeDelivered());
-    $this->assertEqual($message_id, $second_report->getMessageId());
+    $this->assertEquals("status message", $second_report->getStatusMessage());
+    $this->assertEquals($delivered_time, $second_report->getTimeDelivered());
+    $this->assertEquals($message_id, $second_report->getMessageId());
   }
 
 }
