+++ tests/src/Functional/SmsFrameworkMessageTestTrait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\sms\Tests;
+namespace Drupal\Tests\sms\Functional;
 
 use Drupal\sms\Direction;
 use Drupal\sms\Entity\SmsMessageInterface;
